scale factors must be defines as doule
git-svn-id: https://svnsrv.desy.de/desy/h1fitter/trunk@1917 b2a8d302-ca0a-427a-b7b9-081adbce06bc

+++ include/TheorEval.h
@@ -156,8 +156,8 @@ class TheorEval{
  private:
   int _dsId;
   int _iOrd;
-  int _xmur;
-  int _xmuf;
+  double _xmur;
+  double _xmuf;
   int _nTerms;
   double _units;
   vector<string> _termNames;
